added reload function to boot.py so I can reload a module into cache after uploading it via the repl.
Since micro python imports cached modules instead of checking for
changes and loading changed modules, having a reload function will tell
python that we want to load the module again. From this link:
https://forum.micropython.org/viewtopic.php?f=2&t=413&sid=dd97da0f830f213dd9ca6dbf8dcaa50d&start=30
Now once you make a change to a module and upload it to the flash, do a
"reload(module_name)" and then re import that module in order to see the
changes.